backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.2.0
python-decouple==3.8
setuptools
psycopg2-binary>=2.9.0
channels==4.1.0
daphne
